backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.5.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
pytest==7.4.2
pytest-flask==1.2.0
faker==19.6.2
